pyproject.toml
[project]
name = "TMatu"
version = "0.1.0"
description = "Travail de maturité de Chloé"
requires-python = ">=3.13"
dependencies = [
    "clint>=0.5.1",
    "pyaml>=6.0.2",
    "flask>=3.1.2"
]
